requirements.txt
Flask==2.3.2
google-generativeai
Pillow